Add `ToBeFixedForIsolatedProjectsRule`
